add note editor
a unity editor tool
the user can type notes in a text field in the editor windows
the text field can scroll
the editor window can be opened from the menu: Window/Note Editor

notes are text files, saved in a folder, just above the asset folder. Ensure you use a folder relative to the project.
for example:
the assets folder lives in my-project/Assets
the notes folder would be my-project/Notes
do not hardcode my-project, this is an example

notes are named after the asset GUID
assets can be, but not limited too, materials, scripts, prefabs, folders, ...
notes are only added when the user types a note in the editor window. do not create empty notes.
if a user removes all text from a note, delete the note.